Fix instruction bugs, add unit tests, update build and README
- Fix AND/OR/XOR mnemonics (all were incorrectly set to "add")
- Fix BLS to extend BranchInstruction instead of SimpleInstruction
- Fix TOF deserialize to handle unsigned hex byte values (80-FF)
- Add 37 unit tests covering CPU, assembler, memory, and TOF
- Update build.sbt to Scala 3.8.2 with current dependency versions
- Update sbt to 1.10.11, replace sbt-github-packages with sbt-pgp/sonatype
- Write README with architecture docs, ISA reference, and usage examples

+++ src/main/scala/io/github/edadma/trisc/TOF.scala
@@ -78,7 +78,7 @@ object TOF:
         if b.segmentDefined(name) then sys.error(s"duplicate segment on line $l")
         b.segment(name, java.lang.Long.parseUnsignedLong(org, 16))
       case (s"DATA:$data", l) =>
-        b ++= data grouped 2 map (b => java.lang.Byte.parseByte(b, 16))
+        b ++= data grouped 2 map (b => java.lang.Integer.parseInt(b, 16).toByte)
       case (s"RES:$size", l) =>
         b addRes java.lang.Integer.parseUnsignedInt(size, 16)
       case (_, l) => sys.error(s"error on line $l")
